Add vector calibration workbench and PCA estimator support in CAA
- Add aisteer360/workbenches/vector_calibration package with FastAPI-based interface
- Wire ContrastiveDirectionEstimator into CAA for pca_pairwise training method
- Cast pairwise diffs to float before numpy conversion in contrastive estimator
- Add workbenches optional dependency group (fastapi, uvicorn) and seaborn/python-slugify
- Update .gitignore for logs and examples/scripts

+++ aisteer360/algorithms/state_control/common/estimators/contrastive_direction_estimator.py
@@ -241,7 +241,7 @@ class ContrastiveDirectionEstimator(BaseEstimator[SteeringVector]):
             Hn = _pool_over_spans(hs_neg[layer_id], spans_neg)  # [N, H]
 
             # compute pairwise differences
-            diffs = (Hp - Hn).numpy()  # [N, H]
+            diffs = (Hp - Hn).float().numpy()  # [N, H]
 
             if spec.method == "pca_pairwise":
                 # fit PCA to get principal direction
